requirements.txt
Pillow
openai
python-dotenv
customtkinter
